refactor: modularize server and add config
