+++ package.json
@@ -35,7 +35,8 @@
     "pyodide": "^0.28.0",
     "cors": "^2.8.5",
     "swagger-jsdoc": "^6.2.8",
-    "swagger-ui-express": "^5.0.0"
+    "swagger-ui-express": "^5.0.0",
+    "dotenv": "^16.3.1"
   },
   "devDependencies": {
     "nodemon": "^3.1.7",
@@ -55,7 +56,7 @@
   },
   "os": [
     "darwin",
-    "linux", 
+    "linux",
     "win32"
   ],
   "cpu": [
