+++ src/routes/health.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const { health } = require('../controllers/healthController');
+
+const router = express.Router();
+
+router.get('/health', health);
+
+module.exports = router;
