+++ src/routes/upload.js
@@ -1,0 +1,38 @@
+const express = require('express');
+const multer = require('multer');
+const path = require('path');
+const fs = require('fs');
+const config = require('../config');
+const { uploadCsv } = require('../controllers/uploadController');
+
+const router = express.Router();
+
+if (!fs.existsSync(config.uploadDir)) {
+  fs.mkdirSync(config.uploadDir, { recursive: true });
+}
+
+const storage = multer.diskStorage({
+  destination: (req, file, cb) => cb(null, config.uploadDir),
+  filename: (req, file, cb) => {
+    const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1e9);
+    cb(null, file.fieldname + '-' + uniqueSuffix + path.extname(file.originalname));
+  },
+});
+
+const upload = multer({
+  storage,
+  limits: { fileSize: config.maxFileSize, files: 1 },
+  fileFilter: (req, file, cb) => {
+    const allowedTypes = ['.csv', '.json', '.txt', '.py'];
+    const ext = path.extname(file.originalname).toLowerCase();
+    if (allowedTypes.includes(ext)) {
+      cb(null, true);
+    } else {
+      cb(new Error(`File type ${ext} not allowed. Allowed types: ${allowedTypes.join(', ')}`));
+    }
+  },
+});
+
+router.post('/upload-csv', upload.single('csvFile'), uploadCsv);
+
+module.exports = router;
