+++ src/server.js
@@ -1,604 +1,32 @@
-/**
- * Main Express server for Pyodide Python execution
- *
- * This server provides REST API endpoints for executing Python code,
- * uploading files, and managing Python packages using Pyodide.
- *
- * Pyodide 0.28.0 embeds a full CPython 3.13 interpreter compiled to
- * WebAssembly. By loading that runtime inside Node.js we can execute Python
- * code directly from JavaScript without launching a separate Python process.
- * The Express routes below expose this capability as a REST API.
- */
-
-const express = require('express');
-const multer = require('multer');
-const path = require('path');
-const fs = require('fs');
-
-// Import services and utilities
-// pyodideService handles initialization of the WebAssembly runtime and
-// exposes helper methods for executing code inside that runtime.
+const app = require('./app');
+const config = require('./config');
 const pyodideService = require('./services/pyodide-service');
 const logger = require('./utils/logger');
-// Route modules containing all API endpoints for Python execution and file
-// management within Pyodide's virtual filesystem.
-const executeRoutes = require('./routes/execute');
-const fileRoutes = require('./routes/files');  // ← NEW IMPORT
 
-// Import middleware
-const { validateCode, validatePackage } = require('./middleware/validation');
-
-// Import Swagger configuration
-const { swaggerSpec, swaggerUi, swaggerUiOptions } = require('./swagger-config');
-
-// Configuration
-const PORT = process.env.PORT || 3000;
-const MAX_FILE_SIZE = parseInt(process.env.MAX_FILE_SIZE) || 10 * 1024 * 1024; // 10MB
-const UPLOAD_DIR = process.env.UPLOAD_DIR || 'uploads';
-
-// Create Express app
-const app = express();
-
-// Ensure upload directory exists
-if (!fs.existsSync(UPLOAD_DIR)) {
-  fs.mkdirSync(UPLOAD_DIR, { recursive: true });
-}
-
-// Configure multer for file uploads (v2.x syntax)
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, UPLOAD_DIR);
-  },
-  filename: function (req, file, cb) {
-    // Generate unique filename
-    const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1E9);
-    cb(null, file.fieldname + '-' + uniqueSuffix + path.extname(file.originalname));
-  }
-});
-
-const upload = multer({
-  storage: storage,
-  limits: {
-    fileSize: MAX_FILE_SIZE,
-    files: 1
-  },
-  fileFilter: (req, file, cb) => {
-    // Accept CSV, JSON, TXT, and Python files
-    const allowedTypes = ['.csv', '.json', '.txt', '.py'];
-    const ext = path.extname(file.originalname).toLowerCase();
-
-    if (allowedTypes.includes(ext)) {
-      cb(null, true);
-    } else {
-      cb(new Error(`File type ${ext} not allowed. Allowed types: ${allowedTypes.join(', ')}`));
-    }
-  }
-});
-
-// Middleware
-app.use(express.json({ limit: '30mb' }));
-app.use(express.urlencoded({ extended: true, limit: '30mb' }));
-
-// CORS middleware
-app.use((req, res, next) => {
-  res.header('Access-Control-Allow-Origin', process.env.CORS_ORIGIN || '*');
-  res.header('Access-Control-Allow-Methods', 'GET, POST, PUT, DELETE, OPTIONS');
-  res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept, Authorization');
-
-  if (req.method === 'OPTIONS') {
-    res.sendStatus(200);
-  } else {
-    next();
-  }
-});
-
-// Request logging middleware
-app.use((req, res, next) => {
-  logger.info(`${req.method} ${req.path} - ${req.ip}`);
-  next();
-});
-
-// Serve static files from public directory
-app.use(express.static(path.join(__dirname, '../public')));
-
-// Swagger Documentation
-/**
- * @swagger
- * /:
- *   get:
- *     summary: Web interface for testing the API
- *     description: Serves an interactive HTML interface for testing Python code execution
- *     tags: [System]
- *     responses:
- *       200:
- *         description: HTML interface loaded successfully
- *         content:
- *           text/html:
- *             schema:
- *               type: string
- */
-
-// API Documentation endpoint - Swagger UI
-app.use('/docs', swaggerUi.serve);
-app.get('/docs', swaggerUi.setup(swaggerSpec, swaggerUiOptions));
-
-// API Documentation JSON endpoint
-app.get('/docs.json', (req, res) => {
-  res.setHeader('Content-Type', 'application/json');
-  res.send(swaggerSpec);
-});
-
-// Use route modules
-app.use('/api', executeRoutes);  // Python execution routes
-app.use('/api', fileRoutes);     // ← NEW: File management routes
-
-/**
- * @swagger
- * /health:
- *   get:
- *     summary: Main health check endpoint
- *     description: Comprehensive health check including server status, Pyodide status, and system information
- *     tags: [System]
- *     responses:
- *       200:
- *         description: Health check completed successfully
- *         content:
- *           application/json:
- *             schema:
- *               $ref: '#/components/schemas/HealthResponse'
- *       500:
- *         description: Health check failed
- *         content:
- *           application/json:
- *             schema:
- *               $ref: '#/components/schemas/ErrorResponse'
- */
-app.get('/health', (req, res) => {
-  const status = pyodideService.getStatus();
-  const logInfo = logger.getLogInfo();
-
-  res.json({
-    status: 'ok',
-    server: 'running',
-    pyodide: status,
-    logging: logInfo,
-    timestamp: new Date().toISOString(),
-    uptime: process.uptime(),
-    memory: process.memoryUsage()
-  });
-});
-
-// Add this endpoint to your server.js (before error handling middleware)
-
-/**
- * @swagger
- * /api/execute-raw:
- *   post:
- *     summary: Execute Python code from raw text
- *     description: Execute Python code sent as plain text (bypasses JSON escaping issues)
- *     tags: [Python Execution]
- *     requestBody:
- *       required: true
- *       content:
- *         text/plain:
- *           schema:
- *             type: string
- *             example: |
- *               import pandas as pd
- *               import numpy as np
- *               
- *               def analyze_data():
- *                   """
- *                   Triple quoted docstring works fine!
- *                   """
- *                   name = "Alice"
- *                   result = f"Hello {name}!"  # f-strings work!
- *                   return result
- *               
- *               analyze_data()
- *         application/x-python:
- *           schema:
- *             type: string
- *     responses:
- *       200:
- *         description: Code execution completed
- *         content:
- *           application/json:
- *             schema:
- *               $ref: '#/components/schemas/ExecuteResponse'
- *       400:
- *         $ref: '#/components/responses/BadRequest'
- *       500:
- *         $ref: '#/components/responses/InternalError'
- */
-app.post('/api/execute-raw', express.text({ limit: '10mb' }), async (req, res) => {
-  try {
-    const code = req.body;
-
-    if (!code || typeof code !== 'string' || !code.trim()) {
-      return res.status(400).json({
-        success: false,
-        error: 'No Python code provided in request body',
-        timestamp: new Date().toISOString()
-      });
-    }
-
-    logger.info('Executing raw Python code:', {
-      codeLength: code.length,
-      ip: req.ip,
-      contentType: req.get('Content-Type')
-    });
-
-    // Execute the code using the same service
-    const result = await pyodideService.executeCode(code);
-
-    if (result.success) {
-      logger.info('Raw code execution successful');
-    } else {
-      logger.warn('Raw code execution failed:', result.error);
-    }
-
-    res.json(result);
-
-  } catch (error) {
-    logger.error('Raw execution endpoint error:', error);
-    res.status(500).json({
-      success: false,
-      error: error.message,
-      timestamp: new Date().toISOString()
-    });
-  }
-});
-
-// Also add context support for raw text
-app.post('/api/execute-raw-with-context', express.json({ limit: '10mb' }), async (req, res) => {
-  try {
-    const { code, context, timeout } = req.body;
-
-    if (!code || typeof code !== 'string') {
-      return res.status(400).json({
-        success: false,
-        error: 'No Python code provided',
-        timestamp: new Date().toISOString()
-      });
-    }
-
-    logger.info('Executing raw Python code with context:', {
-      codeLength: code.length,
-      hasContext: !!context,
-      timeout: timeout || 'default',
-      ip: req.ip
-    });
-
-    const result = await pyodideService.executeCode(code, context, timeout);
-
-    res.json(result);
-
-  } catch (error) {
-    logger.error('Raw execution with context error:', error);
-    res.status(500).json({
-      success: false,
-      error: error.message,
-      timestamp: new Date().toISOString()
-    });
-  }
-});
-
-/**
- * @swagger
- * /api/upload-csv:
- *   post:
- *     summary: Upload and process CSV file
- *     description: |
- *       Upload a CSV file, load it into Pyodide's virtual filesystem, and return detailed analysis.
- *       
- *       **File Handling:**
- *       - Files are temporarily stored in the uploads folder with unique names
- *       - Files are loaded into Pyodide's virtual filesystem with their original names (sanitized)
- *       - By default, files are kept in the uploads folder (can be configured to auto-delete)
- *       
- *       **Analysis Includes:**
- *       - File dimensions (rows × columns)
- *       - Column names and data types
- *       - Memory usage
- *       - Null value counts
- *       - Sample data (first 3 rows)
- *       - Statistical summary for numeric columns
- *     tags: [File Operations]
- *     requestBody:
- *       required: true
- *       content:
- *         multipart/form-data:
- *           schema:
- *             type: object
- *             properties:
- *               csvFile:
- *                 type: string
- *                 format: binary
- *                 description: CSV file to upload (max 10MB)
- *     responses:
- *       200:
- *         description: File uploaded and processed successfully
- *         content:
- *           application/json:
- *             schema:
- *               $ref: '#/components/schemas/UploadResponse'
- *       400:
- *         $ref: '#/components/responses/BadRequest'
- *       500:
- *         $ref: '#/components/responses/InternalError'
- */
-// Replace your upload endpoint in server.js with this debugging version
-
-app.post('/api/upload-csv', upload.single('csvFile'), async (req, res) => {
-  let tempFilePath = null;
-  
-  try {
-    if (!req.file) {
-      return res.status(400).json({
-        success: false,
-        error: 'No file uploaded'
-      });
-    }
-
-    tempFilePath = req.file.path;
-    
-    logger.info('=== CSV UPLOAD DEBUG START ===');
-    logger.info('File info:', {
-      originalName: req.file.originalname,
-      size: req.file.size,
-      mimetype: req.file.mimetype,
-      tempPath: tempFilePath
-    });
-
-    // Check Pyodide status
-    const pyodideStatus = pyodideService.getStatus();
-    logger.info('Pyodide status:', pyodideStatus);
-    
-    if (!pyodideStatus.isReady) {
-      return res.status(503).json({
-        success: false,
-        error: 'Pyodide is not ready yet',
-        pyodideStatus: pyodideStatus
-      });
-    }
-
-    // Read the uploaded file
-    const fileContent = fs.readFileSync(tempFilePath, 'utf8');
-    logger.info('File content read:', {
-      length: fileContent.length,
-      firstLine: fileContent.split('\n')[0],
-      lineCount: fileContent.split('\n').length
-    });
-
-    // Use original filename (sanitized) in Pyodide
-    const sanitizedName = req.file.originalname.replace(/[^a-zA-Z0-9.-]/g, '_');
-    const pyodideFilename = sanitizedName;
-    logger.info('Using Pyodide filename:', pyodideFilename);
-
-    // Test if we can execute basic Python first
-    logger.info('Testing basic Python execution...');
-    try {
-      const basicTest = await pyodideService.executeCode('2 + 2');
-      logger.info('Basic test result:', basicTest);
-      
-      if (!basicTest.success || basicTest.result !== 4) {
-        throw new Error('Basic Python execution failed');
-      }
-    } catch (basicError) {
-      logger.error('Basic Python test failed:', basicError);
-      return res.status(500).json({
-        success: false,
-        error: 'Pyodide is not working properly',
-        details: basicError.message
-      });
-    }
-
-    // Test pandas availability
-    logger.info('Testing pandas availability...');
-    try {
-      const pandasTest = await pyodideService.executeCode('import pandas as pd; pd.__version__');
-      logger.info('Pandas test result:', pandasTest);
-      
-      if (!pandasTest.success) {
-        throw new Error('Pandas not available');
-      }
-    } catch (pandasError) {
-      logger.error('Pandas test failed:', pandasError);
-      return res.status(500).json({
-        success: false,
-        error: 'Pandas is not available',
-        details: pandasError.message
-      });
-    }
-
-    // Now try to load the CSV file step by step
-    logger.info('Step 1: Writing file to Pyodide filesystem...');
-    try {
-      pyodideService.pyodide.FS.writeFile(pyodideFilename, fileContent);
-      logger.info('File written successfully');
-    } catch (writeError) {
-      logger.error('Failed to write file to Pyodide FS:', writeError);
-      return res.status(500).json({
-        success: false,
-        error: 'Failed to write file to Pyodide filesystem',
-        details: writeError.message
-      });
-    }
-
-    // Step 2: Test if file exists in Pyodide
-    logger.info('Step 2: Checking if file exists in Pyodide...');
-    try {
-      const fileExistsTest = await pyodideService.executeCode(`
-import os
-import json
-exists = os.path.exists('${pyodideFilename}')
-size = os.path.getsize('${pyodideFilename}') if exists else 0
-json.dumps({'exists': exists, 'size': size})
-      `);
-      logger.info('File exists test:', fileExistsTest);
-      
-      if (!fileExistsTest.success) {
-        throw new Error('Failed to check file existence');  
-      }
-      
-      // Parse the JSON result
-      let fileInfo;
-      try {
-        fileInfo = JSON.parse(fileExistsTest.result);
-      } catch (parseError) {
-        throw new Error(`Failed to parse file check result: ${fileExistsTest.result}`);
-      }
-      
-      if (!fileInfo.exists) {
-        throw new Error('File does not exist in Pyodide filesystem');
-      }
-    } catch (existsError) {
-      logger.error('File exists test failed:', existsError);
-      return res.status(500).json({
-        success: false,
-        error: 'File was not written to Pyodide filesystem correctly',
-        details: existsError.message
-      });
-    }
-
-    // Step 3: Since the file is confirmed to exist, create a simple analysis
-    logger.info('Step 3: Creating analysis for uploaded file...');
-    try {
-      // Since you mentioned that Python code can open the file correctly,
-      // let's just create a basic analysis based on what we know
-      const analysis = {
-        success: true,
-        shape: [0, 0], // We'll let the user's Python code determine this
-        columns: [],   // We'll let the user's Python code determine this
-        has_data: true,
-        memory_usage: req.file.size,
-        message: 'File uploaded successfully. Use Python code to analyze the data.'
-      };
-      
-      logger.info('Analysis created:', analysis);
-      logger.info('=== CSV UPLOAD DEBUG END - SUCCESS ===');
-      
-      res.json({
-        success: true,
-        file: {
-          originalName: req.file.originalname,
-          size: req.file.size,
-          pyodideFilename: pyodideFilename,
-          tempPath: tempFilePath,
-          keepFile: true
-        },
-        analysis: analysis,
-        debug: {
-          pyodideStatus: pyodideStatus,
-          fileWritten: true,
-          pandasAvailable: true,
-          csvLoaded: true,
-          note: 'File is available in Pyodide filesystem as: ' + pyodideFilename
-        },
-        timestamp: new Date().toISOString()
-      });
-      
-    } catch (error) {
-      logger.error('Analysis creation failed:', error);
-      return res.status(500).json({
-        success: false,
-        error: 'Failed to create file analysis',
-        details: error.message
-      });
-    }
-
-  } catch (error) {
-    logger.error('=== CSV UPLOAD DEBUG END - ERROR ===');
-    logger.error('Upload error:', error);
-    logger.error('Error stack:', error.stack);
-
-    // Clean up on error
-    if (tempFilePath && fs.existsSync(tempFilePath)) {
-      try {
-        fs.unlinkSync(tempFilePath);
-        logger.info('Cleaned up temp file on error');
-      } catch (cleanupError) {
-        logger.warn('Failed to cleanup temp file:', cleanupError.message);
-      }
-    }
-
-    res.status(500).json({
-      success: false,
-      error: error.message,
-      stack: process.env.NODE_ENV === 'development' ? error.stack : undefined,
-      timestamp: new Date().toISOString()
-    });
-  }
-});
-
-// If you want "/" to always serve your custom index.html:
-app.get('/', (req, res) => {
-  res.sendFile(path.join(__dirname, '..', 'public', 'index.html'));
-});
-
-
-// Error handling middleware
-app.use((err, req, res, next) => {
-  logger.error('Unhandled error:', err);
-
-  if (err instanceof multer.MulterError) {
-    if (err.code === 'LIMIT_FILE_SIZE') {
-      return res.status(400).json({
-        success: false,
-        error: `File size too large. Maximum size: ${MAX_FILE_SIZE / 1024 / 1024}MB`
-      });
-    }
-  }
-
-  res.status(500).json({
-    success: false,
-    error: process.env.NODE_ENV === 'production' ? 'Internal server error' : err.message,
-    timestamp: new Date().toISOString()
-  });
-});
-
-// 404 handler
-app.use((req, res) => {
-  res.status(404).json({
-    success: false,
-    error: 'Endpoint not found',
-    timestamp: new Date().toISOString()
-  });
-});
-
-/**
- * Initialize and start the server
- */
 async function startServer() {
   try {
     logger.info('Starting Pyodide Express Server...');
-
-    // Initialize Pyodide
     logger.info('Initializing Pyodide...');
     await pyodideService.initialize();
     logger.info('Pyodide initialization completed!');
 
-    // Start the Express server
-    const server = app.listen(PORT, () => {
+    const server = app.listen(config.port, () => {
       const logInfo = logger.getLogInfo();
-
-      logger.info(`🚀 Server running on port ${PORT}`);
-      logger.info(`📖 Web interface: http://localhost:${PORT}`);
-      logger.info(`📚 API Documentation: http://localhost:${PORT}/docs`);
-      logger.info(`🔧 API base URL: http://localhost:${PORT}/api`);
-      logger.info(`📊 Health check: http://localhost:${PORT}/health`);
-      logger.info(`📁 File management: http://localhost:${PORT}/api/uploaded-files`);
+      logger.info(`🚀 Server running on port ${config.port}`);
+      logger.info(`📖 Web interface: http://localhost:${config.port}`);
+      logger.info(`📚 API Documentation: http://localhost:${config.port}/docs`);
+      logger.info(`🔧 API base URL: http://localhost:${config.port}/api`);
+      logger.info(`📊 Health check: http://localhost:${config.port}/health`);
+      logger.info(`📁 File management: http://localhost:${config.port}/api/uploaded-files`);
 
       if (logInfo.isFileLoggingEnabled) {
         logger.info(`📝 Logs writing to: ${logInfo.logFile}`);
         logger.info(`📁 Log directory: ${logInfo.logDirectory}`);
       } else {
-        logger.info(`📺 Console-only logging (no file output)`);
+        logger.info('📺 Console-only logging (no file output)');
       }
     });
 
-    // Graceful shutdown
     process.on('SIGTERM', () => {
       logger.info('SIGTERM received, shutting down gracefully...');
       server.close(() => {
@@ -616,14 +44,12 @@ async function startServer() {
     });
 
     return server;
-
   } catch (error) {
     logger.error('Failed to start server:', error);
     process.exit(1);
   }
 }
 
-// Start the server if this file is run directly
 if (require.main === module) {
   startServer();
 }
